Fixed merge issues and restructured free eval stuff
